+++ backend/alembic/versions/2022-11-16-09-11-04_8e63db7aacb7_create_data_table.py
@@ -22,7 +22,7 @@ def upgrade() -> None:
         'data',
         sa.Column('id', sa.Integer(), nullable=False),
         sa.Column('name', sa.String()),
-        sa.Column('form', sa.Integer(), sa.ForeignKey('form.id')),
+        sa.Column('form', sa.BigInteger(), sa.ForeignKey('form.id')),
         sa.Column('geo', pg.ARRAY(sa.Float()), nullable=True),
         sa.Column(
             'created', sa.DateTime(), nullable=True,
